extract closure-kinds query to separate file
